requirements.txt
Django>=4.2.0
django-cors-headers>=4.3.0
django-environ>=0.11.2
django-filter>=23.3
django-rest-framework>=0.1.0
django-rest-swagger>=2.2.0
drf-yasg>=1.21.7
psycopg2>=2.9.9
Pillow>=10.1.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.1.0